Extract consolidate helpers from index.ts into tools/consolidate-helpers.ts
Move 8 consolidate-related helper functions (detectDuplicates, findClusters,
suggestMerges, executeMerge, pruneSuperseded, dryRunAll, plus internal
loadEmbeddingsByNoteId and findExistingExecuteMergeTarget) from src/index.ts
into a new src/tools/consolidate-helpers.ts module.

Follows the established extraction pattern from helpers/project.ts where
ServerContext (ctx) is passed as an explicit parameter to functions that need
it (executeMerge, pruneSuperseded), rather than using module-level closures.

Pure analysis functions (detectDuplicates, findClusters, suggestMerges,
dryRunAll) don't require ctx access.

Also cleans up unused imports from index.ts: classifyTheme, recencyScore,
ConsolidateResult, ConsolidateExecuteMergeEvidence.

Reduces index.ts by ~944 lines.
